Update and rename NN Layout.py to nn-layout.py
Aligned code to mostly match pylint specifications
